Rewrite app_list_page in TS
